# doc/api.yaml
openapi: 3.0.1
info:
  title: mcs api
  description: list of main api
  version: 1.0.0
servers:
- url: https://mcs.cloud273.com/app/api
paths:
...
  /doctor:
    get:
      tags:
      - doctor/information
      summary: Get doctor account
      operationId: getDoctorAccount
      parameters:
      - name: os
        in: header
        description: Device OS
        required: true
        schema:
          type: string
          default: ios
      - name: app
        in: header
        description: Application identifier
        required: true
        schema:
          type: string
          default: "com.cloud273.doctor|1.0.0"
      - name: token
        in: header
        description: Access token
        required: true
        schema:
          type: string
          default: "doctor_token_here"
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DoctorDetailWithCert'   
        400:
          description: Bad request
        403:
          description: Invalid token
        500:
          description: Unknown
...
  /doctor/working-time/list:
    get:
      tags:
      - doctor/information
      summary: Get list working time of doctor
      operationId: getListWorkingTimeByDoctor
      parameters:
      - name: os
        in: header
        description: Device OS
        required: true
        schema:
          type: string
          default: ios
      - name: app
        in: header
        description: Application identifier
        required: true
        schema:
          type: string
          default: "com.cloud273.doctor|1.0.0"
      - name: token
        in: header
        description: Access token
        required: true
        schema:
          type: string
          default: "doctor_token_here"
      - name: from
        in: query
        description: from
        required: true
        schema:
          type: string
          default: "2019-12-12T07%3A00%3A00%2B07%3A00"
      - name: to
        in: query
        description: to
        required: true
        schema:
          type: string
          default: "2020-01-31T07%3A00%3A00%2B07%3A00"
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkingTimeDetail'   
        400:
          description: Bad request
        403:
          description: Invalid token
        500:
          description: Unknown
...
  /patient/booking/specialty/list:
    get:
      tags:
      - patient/booking
      summary: Get list specialty
      operationId: getListActiveSpecialtyByPatient
      parameters:
      - name: os
        in: header
        description: Device OS
        required: true
        schema:
          type: string
          default: ios
      - name: app
        in: header
        description: Application identifier
        required: true
        schema:
          type: string
          default: "com.cloud273.patient|1.0.0"
      - name: token
        in: header
        description: Access token
        required: true
        schema:
          type: string
          default: "patient_token_here"
      - name: symptoms
        in: query
        description: symptoms
        required: true
        schema:
          type: string
          default: "base64 of json serialize from array of symptom"
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SpecialtyConfig'
        400:
          description: Bad request
        403:
          description: Invalid token
        500:
          description: Unknown
  /patient/booking/doctor/list:
    get:
      tags:
      - patient/booking
      summary: Get list doctor
      operationId: getListActiveDoctorByPatient
      parameters:
      - name: os
        in: header
        description: Device OS
        required: true
        schema:
          type: string
          default: ios
      - name: app
        in: header
        description: Application identifier
        required: true
        schema:
          type: string
          default: "com.cloud273.patient|1.0.0"
      - name: token
        in: header
        description: Access token
        required: true
        schema:
          type: string
          default: "patient_token_here"
      - name: specialty
        in: query
        description: specialty
        required: true
        schema:
          type: string
          default: "primary"
          enum: ["primary", "cardiology"]
      - name: type
        in: query
        description: type
        required: true
        schema:
          type: string
          default: "classic"
          enum: ["classic", "telemed"]
      - name: longitude
        in: query
        description: longitude
        schema:
          type: number
          format: float
          default: 10.001
      - name: latitude
        in: query
        description: latitude
        schema:
          type: number
          format: float
          default: 10.001
      - name: offsetName
        in: query
        description: offsetName
        schema:
          type: string
          default: name
      - name: limit
        in: query
        description: limit
        schema:
          type: number
          default: 50
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DoctorBasic'
        400:
          description: Bad request
        403:
          description: Invalid token
        500:
          description: Unknown
  /patient/booking/doctor:
    get:
      tags:
      - patient/booking
      summary: Get doctor
      operationId: getActiveDoctorByPatient
      parameters:
      - name: os
        in: header
        description: Device OS
        required: true
        schema:
          type: string
          default: ios
      - name: app
        in: header
        description: Application identifier
        required: true
        schema:
          type: string
          default: "com.cloud273.patient|1.0.0"
      - name: token
        in: header
        description: Access token
        required: true
        schema:
          type: string
          default: "patient_token_here"
      - name: id
        in: query
        description: id
        required: true
        schema:
          type: number
          default: 1
      - name: specialty
        in: query
        description: specialty
        required: true
        schema:
          type: string
          default: "primary"
          enum: ["primary", "cardiology"]
      - name: type
        in: query
        description: type
        required: true
        schema:
          type: string
          default: "classic"
          enum: ["classic", "telemed"]
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DoctorDetailWithCertAndPackage'
        400:
          description: Bad request
        403:
          description: Invalid token
        404:
          description: Not found
        500:
          description: Unknown
  /patient/booking/working-time/list:
    get:
      tags:
      - patient/booking
      summary: Get working time of package
      operationId: getWorkingTimeByPatient
      parameters:
      - name: os
        in: header
        description: Device OS
        required: true
        schema:
          type: string
          default: ios
      - name: app
        in: header
        description: Application identifier
        required: true
        schema:
          type: string
          default: "com.cloud273.patient|1.0.0"
      - name: token
        in: header
        description: Access token
        required: true
        schema:
          type: string
          default: "patient_token_here"
      - name: packageId
        in: query
        description: packageId
        required: true
        schema:
          type: number
          default: 1
      - name: from
        in: query
        description: from
        required: true
        schema:
          type: string
          default: "2019-12-12T07%3A00%3A00%2B07%3A00"
      - name: to
        in: query
        description: to
        required: true
        schema:
          type: string
          default: "2020-01-31T07%3A00%3A00%2B07%3A00"
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkingTimeDetail'
        400:
          description: Bad request
        403:
          description: Invalid token
        404:
          description: Not found
        500:
          description: Unknown
  /patient/booking/time/list:
    get:
      tags:
      - patient/booking
      summary: Get booking time of package
      operationId: getBookingTimeByPatient
      parameters:
      - name: os
        in: header
        description: Device OS
        required: true
        schema:
          type: string
          default: ios
      - name: app
        in: header
        description: Application identifier
        required: true
        schema:
          type: string
          default: "com.cloud273.patient|1.0.0"
      - name: token
        in: header
        description: Access token
        required: true
        schema:
          type: string
          default: "patient_token_here"
      - name: packageId
        in: query
        description: packageId
        required: true
        schema:
          type: number
          default: 1
      - name: from
        in: query
        description: from
        required: true
        schema:
          type: string
          default: "2020-01-01T07%3A30%3A00%2B07%3A00"
      - name: to
        in: query
        description: to
        required: true
        schema:
          type: string
          default: "2020-01-01T12%3A00%3A00%2B07%3A00"
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DateTime'
        400:
          description: Bad request
        403:
          description: Invalid token
        404:
          description: Not found
        500:
          description: Unknown
...
components:
  schemas:
...
    Dob:
      type: string
      format: yyyy-MM-dd
      example: "2000-01-01"
      default: "2000-01-01"
    Bool:
      type: boolean
      example: true
      default: true
...
    CurrencyUnit:
      type: string
      example: vnd
      default: vnd
      enum: [vnd, usd]
    Date:
      type: string
      format: yyyy-MM-dd
      example: "2020-01-01"
      default: "2020-01-01"
    DateTime:
      type: string
      format: yyyy-MM-ddTHH:mm:ssZ
      example: "2020-01-01T08:00:00+07:00"
      default: "2020-01-01T08:00:00+07:00"
...
    DoctorCertType:
      type: string
      example: working
      default: working
      enum: [personal, working, degree, other]
...
    Firstname:
      maxLength: 128
      minLength: 2
      type: string
      example: "Dung"
      default: "Dung"
    Gender:
      type: string
      example: male
      default: male
      enum: [male, female]
    ID:
      type: integer
      format: int64
      example: 1
      default: 1
    ImageUrl:
      maxLength: 128
      minLength: 4
      type: string
      example: https://cloud273.com/mcs/resource/image.png
      default: https://cloud273.com/mcs/resource/image.png
    Language:
      type: string
      example: vi
      default: vi
      enum: [vi, en]
    Lastname:
      maxLength: 128
      type: string
      example: "Nguyen"
      default: "Nguyen"
...
    Description:
      maxLength: 1000
      type: string
      example: "Description"
      default: "Description"
    Number:
      type: number
      example: 10
      default: 10
    OtherDate:
      type: string
      format: yyyy-MM-dd
      example: "2020-01-31"
      default: "2020-01-31"
...
    PackageType:
      type: string
      example: "classic"
      default: "classic"
      enum: ["classic", "telemed"]
...
    SpecialtyCode:
      type: string
      example: "primary"
      default: "primary"
      enum: ["primary", "cardiology"]
...
    Timezone:
      type: string
      example: "+07:00"
      default: "+07:00"
...
    Username:
      maxLength: 128
      type: string
      example: "nglequduph@gmail.com"
      default: "nglequduph@gmail.com"
      description: "Phone or email"
    VisitTime:
      type: number
      example: 600
      default: 600
...
    Certificate:
      type: object
      required: [code, name, issuer, issueDate, image]
      properties:
        code:
          type: string
          minLength: 6
          maxLength: 32
          example: "100000"
          default: "100000"
        name:
          type: string
          minLength: 6
          maxLength: 64
          example: Certificate of
          default: Certificate of
        issuer:
          type: string
          minLength: 2
          maxLength: 64
          example: SYT HCM
          default: SYT HCM
        issueDate:
          $ref: '#/components/schemas/Date'
        expDate:
          $ref: '#/components/schemas/OtherDate'
        image:
          $ref: '#/components/schemas/ImageUrl'
...
    Price:
      type: object
      required: [amount, currency]
      properties:
        amount:
          $ref: '#/components/schemas/Number'
        currency:
          $ref: '#/components/schemas/CurrencyUnit'
    Profile:
      type: object
      required: [firstname, lastname, gender, dob]
      properties:
        firstname:
          $ref: '#/components/schemas/Firstname'
        lastname:
          $ref: '#/components/schemas/Lastname'
        gender:
          $ref: '#/components/schemas/Gender'
        dob:
          $ref: '#/components/schemas/Dob'      
...
    SpecialtyConfig:
      type: object
      required: [code, image, name]
      properties:
        code:
          $ref: '#/components/schemas/SpecialtyCode'
        image:
          $ref: '#/components/schemas/ImageUrl'
        name:
          $ref: '#/components/schemas/LocalizeText'
...
    LocalizeText:
      type: object
      required: [vi, en]
      properties:
        vi:
          type: string
          example: vietnamese text
          default: vietnamese text 
        en:
          type: string
          example: english text
          default: english text 
...
    DoctorBasic:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/ID'
        image:
          $ref: '#/components/schemas/ImageUrl'
        profile:
          $ref: '#/components/schemas/Profile'
        specialties:
          type: array
          items:
            $ref: '#/components/schemas/SpecialtyCode'
        title:
          type: string
          maxLength: 32
          default: Dr
          example: Dr
        office:
          type: string
          maxLength: 256
          default: CEO Hospital
          example: CEO Hospital
        startWork:
          $ref: '#/components/schemas/Date'
    DoctorDetail:
      allOf:
        - $ref: '#/components/schemas/DoctorBasic'
      type: object
      properties:
        language:
          $ref: '#/components/schemas/Language'
        username:
          $ref: '#/components/schemas/Username'
        specialties:
          type: array
          items:
            $ref: '#/components/schemas/SpecialtyCode'
        biography:
          type: string
          maxLength: 10000
          default: Something about doctor's biography
          example: Something about doctor's biography
        timezone:
          $ref: '#/components/schemas/Timezone'
        description:
          $ref: '#/components/schemas/Description'
        deactivated:
          $ref: '#/components/schemas/Bool'
    DoctorDetailWithCert:
      allOf:
        - $ref: '#/components/schemas/DoctorDetail'
      type: object
      properties:
        certificates:
          type: array
          items:
            $ref: '#/components/schemas/DoctorCertDetail'
    DoctorDetailWithCertAndPackage:
      allOf:
        - $ref: '#/components/schemas/DoctorDetailWithCert'
      type: object
      properties:
        packages:
          type: array
          items:
            $ref: '#/components/schemas/PackageDetail'
...
    DoctorCertUpdateWithID:
      allOf:
        - $ref: '#/components/schemas/Certificate'
      type: object
      required: [type, id]
      properties:
        type:
          $ref: '#/components/schemas/DoctorCertType'
        id:
          $ref: '#/components/schemas/ID'
        description:
          $ref: '#/components/schemas/Description'
    DoctorCertBasic:
      allOf:
        - $ref: '#/components/schemas/DoctorCertUpdateWithID'
    DoctorCertDetail:
      allOf:
        - $ref: '#/components/schemas/DoctorCertBasic'
...
    PackageBasic:
      type: object
      required: [id, specialty, type, price, visitTime]
      properties:
        id:
          $ref: '#/components/schemas/ID'
        specialty:
          $ref: '#/components/schemas/SpecialtyCode'
        type:
          $ref: '#/components/schemas/PackageType'
        price:
          $ref: '#/components/schemas/Price'
        visitTime:
          $ref: '#/components/schemas/VisitTime'
        description:
          $ref: '#/components/schemas/Description'
    PackageDetail:
      allOf:
        - $ref: '#/components/schemas/PackageBasic'
...
    WorkingTimeDetail:
      type: object
      required: [begin, end]
      properties:
        begin:
          $ref: '#/components/schemas/DateTime'
        end:
          $ref: '#/components/schemas/DateTime'
        packageId:
          $ref: '#/components/schemas/Number'
        scheduleId:
          $ref: '#/components/schemas/Number'
        workingDayId:
          $ref: '#/components/schemas/Number'
        visitTime:
          $ref: '#/components/schemas/Number'